refactor: combine advanced search components
