+++ frontend/src/components/AdvancedSearch/index.ts
@@ -1,0 +1,6 @@
+import AdvancedSearch from "./AdvancedSearch.vue";
+import TimeRange from "./TimeRange.vue";
+
+export default AdvancedSearch;
+
+export { TimeRange };
