+++ frontend/src/components/AlterSchemaPrepForm/AlterSchemaPrepForm.vue
@@ -137,11 +137,11 @@
             <NTabPane :tab="$t('common.database')" name="DATABASE">
               <div class="space-y-3">
                 <div class="w-full flex items-center space-x-2">
-                  <AdvancedSearchBox
+                  <AdvancedSearch
                     v-model:params="state.params"
                     :autofocus="false"
                     :placeholder="$t('database.filter-database')"
-                    :support-option-id-list="supportOptionIdList"
+                    :scope-options="scopeOptions"
                   />
                   <DatabaseLabelFilter
                     v-model:selected="state.selectedLabels"
@@ -164,11 +164,11 @@
             </NTabPane>
             <NTabPane :tab="renderDatabaseGroupTabTitle" name="DATABASE_GROUP">
               <div class="space-y-3">
-                <AdvancedSearchBox
+                <AdvancedSearch
                   v-model:params="state.params"
                   :autofocus="false"
                   :placeholder="$t('database.filter-database')"
-                  :support-option-id-list="supportOptionIdList"
+                  :scope-options="scopeOptions"
                 />
                 <DatabaseGroupDataTable
                   :database-group-list="filteredDatabaseGroupList"
@@ -296,7 +296,7 @@ import type {
 import { UNKNOWN_ID, DEFAULT_PROJECT_V1_NAME } from "@/types";
 import { State } from "@/types/proto/v1/common";
 import { TenantMode } from "@/types/proto/v1/project_service";
-import type { SearchScopeId, SearchParams } from "@/utils";
+import type { SearchParams } from "@/utils";
 import {
   allowUsingSchemaEditorV1,
   instanceV1HasAlterSchema,
@@ -307,6 +307,8 @@ import {
   extractInstanceResourceName,
   extractProjectResourceName,
 } from "@/utils";
+import AdvancedSearch from "../AdvancedSearch";
+import { useCommonSearchScopeOptions } from "../AdvancedSearch/useCommonSearchScopeOptions";
 import { DatabaseLabelFilter, DrawerContent } from "../v2";
 import DatabaseGroupPrevEditorModal from "./DatabaseGroupPrevEditorModal.vue";
 import ProjectStandardView from "./ProjectStandardView.vue";
@@ -373,6 +375,15 @@ const state = reactive<LocalState>({
     scopes: [],
   },
 });
+
+const scopeOptions = useCommonSearchScopeOptions(
+  computed(() => state.params),
+  computed(() =>
+    state.databaseSelectedTab === "DATABASE"
+      ? ["project", "instance", "environment"]
+      : ["project", "environment"]
+  )
+);
 
 const hasDatabaseGroupFeature = computed(() => {
   return hasFeature("bb.feature.database-grouping");
@@ -737,12 +748,4 @@ const generateTenant = async () => {
 const cancel = () => {
   emit("dismiss");
 };
-
-const supportOptionIdList = computed((): SearchScopeId[] => {
-  if (state.databaseSelectedTab === "DATABASE") {
-    return ["project", "instance", "environment"];
-  }
-
-  return ["project", "environment"];
-});
 </script>
