+++ frontend/src/components/AuditLog/AuditLogSearch/index.ts
@@ -1,5 +1,3 @@
-import AdvancedSearchBox from "./AdvancedSearchBox.vue";
 import AuditLogSearch from "./AuditLogSearch.vue";
 
 export default AuditLogSearch;
-export { AuditLogSearch, AdvancedSearchBox };
