+++ frontend/src/components/IssueV1/components/IssueSearch/IssueSearch.vue
@@ -4,21 +4,25 @@
       v-if="components.includes('searchbox')"
       class="flex flex-row items-center gap-x-2"
     >
-      <AdvancedSearchBox
-        :params="params"
-        :autofocus="autofocus"
-        :readonly-scopes="readonlyScopes"
-        :support-option-id-list="supportOptionIdList"
+      <AdvancedSearch
         class="flex-1"
-        v-bind="componentProps?.searchbox"
+        :params="params"
+        :readonly-scopes="readonlyScopes"
+        :scope-options="scopeOptions"
         @update:params="$emit('update:params', $event)"
-        @select-unsupported-scope="handleSelectScope"
+      />
+      <TimeRange
+        v-if="components.includes('time-range')"
+        v-model:show="showTimeRange"
+        :params="params"
+        v-bind="componentProps?.['time-range']"
+        @update:params="$emit('update:params', $event)"
       />
       <slot name="searchbox-suffix" />
     </div>
     <slot name="default" />
 
-    <div class="flex flex-col md:flex-row md:items-center gap-y-1">
+    <div class="mt-2 flex flex-col md:flex-row md:items-center gap-y-1">
       <div class="flex-1 flex items-start">
         <Status
           v-if="components.includes('status')"
@@ -28,29 +32,18 @@
         />
         <slot name="primary" />
       </div>
-      <div class="flex flex-row space-x-4">
-        <NInputGroup>
-          <TimeRange
-            v-if="components.includes('time-range')"
-            v-model:show="showTimeRange"
-            :params="params"
-            v-bind="componentProps?.['time-range']"
-            @update:params="$emit('update:params', $event)"
-          />
-          <slot name="secondary" />
-        </NInputGroup>
-      </div>
     </div>
   </div>
 </template>
 
 <script lang="ts" setup>
 import { ref, computed } from "vue";
-import type { SearchParams, SearchScope, SearchScopeId } from "@/utils";
+import AdvancedSearch from "@/components/AdvancedSearch";
+import TimeRange from "@/components/AdvancedSearch/TimeRange.vue";
+import type { SearchParams, SearchScope } from "@/utils";
 import { UIIssueFilterScopeIdList, SearchScopeIdList } from "@/utils";
-import AdvancedSearchBox from "./AdvancedSearchBox.vue";
 import Status from "./Status.vue";
-import TimeRange from "./TimeRange.vue";
+import { useIssueSearchScopeOptions } from "./useIssueSearchScopeOptions";
 
 export type SearchComponent =
   | "searchbox"
@@ -65,7 +58,7 @@ export type SearchComponent =
   | "approver"
   | "approval";
 
-withDefaults(
+const props = withDefaults(
   defineProps<{
     params: SearchParams;
     readonlyScopes?: SearchScope[];
@@ -75,7 +68,7 @@ withDefaults(
   }>(),
   {
     readonlyScopes: () => [],
-    components: () => ["searchbox", "status", "time-range"],
+    components: () => ["searchbox", "time-range", "status"],
     componentProps: undefined,
   }
 );
@@ -85,14 +78,8 @@ defineEmits<{
 
 const showTimeRange = ref(false);
 
-const handleSelectScope = (id: SearchScopeId) => {
-  if (id === "created") {
-    showTimeRange.value = true;
-  }
-};
-
-const supportOptionIdList = computed(() => [
-  ...UIIssueFilterScopeIdList,
-  ...SearchScopeIdList,
-]);
+const scopeOptions = useIssueSearchScopeOptions(
+  computed(() => props.params),
+  computed(() => [...UIIssueFilterScopeIdList, ...SearchScopeIdList])
+);
 </script>
