+++ frontend/src/components/IssueV1/components/IssueSearch/index.ts
@@ -1,5 +1,3 @@
-import AdvancedSearchBox from "./AdvancedSearchBox.vue";
 import IssueSearch from "./IssueSearch.vue";
 
 export default IssueSearch;
-export { IssueSearch, AdvancedSearchBox };
